[COZY-311] refactor: 테스트 후 필요 파일 Refactor
