+++ src/main/java/com/cozymate/cozymate_server/domain/memberstat/dto/MemberStatDifferenceResponseDTO.java
@@ -1,6 +1,5 @@
 package com.cozymate.cozymate_server.domain.memberstat.dto;
 
-import com.cozymate.cozymate_server.domain.room.enums.DifferenceStatus;
 import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
